Update README.md
How it works:
There are 3 multipliers which all are convars, meaning this mod edit is custom and can be effectively disabled.
vfire_pc_spread_base (0.8) : This is the base rate fire will spread to any players. 20% is not a huge reduction, but there's a slight difference and the other multipliers are far greater.
vfire_pc_spread_onfire (0.5) : The rate fire will spread to a player if they are already on fire. This makes the fire more lenient without being too generous.
vfire_pc_spread_self (0.0) : The rate fire on the player another part of the player. I recommend extremely low values if you change this, otherwise it's possible you'll just burn to death like you would in vanilla.

I will not be updating this because i plan on working more on vFire and making my own addon which includes this, also if vFire updates this will probably break, maybe i'll do something about it but no promises